requirements.txt
alembic==1.16.5
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.2.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.1.8
cloudpickle==3.1.2
comm==0.2.3
contourpy==1.3.0
cryptography==46.0.5
cycler==0.12.1
databricks-sdk==0.97.0
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
docker==7.1.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.128.8
fastjsonschema==2.21.2
Flask==3.1.3
fonttools==4.60.2
fqdn==1.5.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.0
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
ipykernel==6.31.0
ipython==8.18.1
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.7
lark==1.3.1
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.9.4
matplotlib-inline==0.2.1
mistune==3.2.0
mlflow==3.1.4
mlflow-skinny==3.1.4
narwhals==2.17.0
nbclient==0.10.2
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.5.4
notebook_shim==0.2.4
numpy==2.0.2
openpyxl==3.1.5
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.6
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
Send2Trash==2.1.0
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
SQLAlchemy==2.0.48
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.49.3
streamlit==1.50.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.4.0
tornado==6.5.4
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
uvicorn==0.39.0
wcwidth==0.6.0
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.6
widgetsnbextension==4.0.15
zipp==3.23.0
